+++ file/src/bring/mod.rs
@@ -15,6 +15,8 @@ use crate::{
 
 use clap::Parser;
 
+use clap_complete::ArgValueCompleter;
+use xvc_core::util::completer::strum_variants_completer;
 use xvc_core::{ContentDigest, RecheckMethod, XvcCachePath, XvcFileType, XvcMetadata, XvcRoot};
 use xvc_ecs::{HStore, XvcStore};
 use xvc_logging::{debug, error, uwr, warn, XvcOutputSender};
@@ -31,6 +33,8 @@ use xvc_walker::PathSync;
 #[command(rename_all = "kebab-case")]
 pub struct BringCLI {
     /// Storage name or guid to send the files
+    ///
+    /// TODO: Add a storage_identifier completer
     #[arg(long, short, alias = "from")]
     storage: StorageIdentifier,
 
@@ -47,10 +51,12 @@ pub struct BringCLI {
 
     /// Recheck (checkout) the file in one of the four alternative ways.
     /// (See `xvc file recheck`) and [RecheckMethod]
-    #[arg(long, alias = "as")]
+    #[arg(long, alias = "as", add = ArgValueCompleter::new(strum_variants_completer::<RecheckMethod>))]
     recheck_as: Option<RecheckMethod>,
 
     /// Targets to bring from the storage
+    ///
+    /// TODO: Add a tracked_targets completer
     #[arg()]
     targets: Option<Vec<String>>,
 }