+++ file/src/hash/mod.rs
@@ -37,10 +37,11 @@ pub struct HashCLI {
     text_or_binary: TextOrBinary,
 
     /// Files to process
+    ///
+    /// NOTE: This uses the default completion as the command can work anywhere with any file
     #[arg()]
     targets: Vec<PathBuf>,
 }
-
 
 impl UpdateFromXvcConfig for HashCLI {
     fn update_from_conf(self, conf: &XvcConfig) -> xvc_config::error::Result<Box<Self>> {