+++ pipeline/src/pipeline/api/export.rs
@@ -3,7 +3,6 @@ use crate::error::{Error, Result};
 use clap::Parser;
 use itertools::Itertools;
 use std::{fs, path::PathBuf};
-use xvc_config::FromConfigKey;
 
 use xvc_core::{
     util::serde::{to_json, to_yaml},
@@ -21,10 +20,6 @@ use crate::{
 #[derive(Debug, Clone, Parser)]
 #[command(name = "export")]
 pub struct ExportCLI {
-    /// Name of the pipeline to export
-    #[arg(long, short)]
-    pipeline_name: Option<String>,
-
     /// File to write the pipeline. Writes to stdout if not set.
     #[arg(long)]
     file: Option<PathBuf>,
@@ -32,6 +27,8 @@ pub struct ExportCLI {
     /// Output format. One of json or yaml. If not set, the format is
     /// guessed from the file extension. If the file extension is not set,
     /// json is used as default.
+    ///
+    /// TODO: Add xvc_pipeline_schema_format_completer
     #[arg(long)]
     format: Option<XvcSchemaSerializationFormat>,
 }
@@ -42,13 +39,18 @@ pub struct ExportCLI {
 /// If `file` is None, prints to stdout.
 /// If `name` is None, uses the default pipeline name from the config.
 /// If `format` is None, uses the default format from [XvcSchemaSerializationFormat::default()]
-pub fn cmd_export(output_snd: &XvcOutputSender, xvc_root: &XvcRoot, opts: ExportCLI) -> Result<()> {
-    let pipeline = XvcPipeline::from_conf(xvc_root.config());
-    let name = opts.pipeline_name.unwrap_or(pipeline.name);
+pub fn cmd_export(
+    output_snd: &XvcOutputSender,
+    xvc_root: &XvcRoot,
+    pipeline_name: &str,
+    opts: ExportCLI,
+) -> Result<()> {
+    let name = pipeline_name;
     let file = opts.file;
     let format = opts.format;
-    let mut p_res: Result<(XvcEntity, XvcPipeline)> =
-        Err(Error::CannotFindPipeline { name: name.clone() });
+    let mut p_res: Result<(XvcEntity, XvcPipeline)> = Err(Error::CannotFindPipeline {
+        name: name.to_owned(),
+    });
 
     xvc_root.with_store(|bs: &XvcStore<XvcPipeline>| {
         let name = name.clone();